Remove one e2e test, increase timeout, to reduce CI flakiness (#811)
* Use the same style as rest of file

* Remove test that is part of the other one anyway

* Increase timeout
